requirements.txt
django
djangorestframework
requests
drf-yasg
